@@ -1,6 +1,6 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { map, mergeMap, takeUntil } from 'rxjs/operators';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { combineLatest, of, Subject } from 'rxjs';
 import { MatBottomSheet } from '@angular/material/bottom-sheet';
@@ -30,7 +30,7 @@ export class AdminComponent implements OnInit, OnDestroy {
   public flightsData: Flight[] = [];
   public passengersData: Passenger[] = [];
   public servicesData: AncillaryService[] = [];
-  public flight: FormControl =  new FormControl('');
+  public flight: UntypedFormControl =  new UntypedFormControl('');
   public passengerTable: SmartTableHeader[] = [
     {
       key: 'full_name',
